refactor #66: MenuAddModal 디렉토리 이동
- Main 디렉토리 안으로 이동
