+++ fe/src/components/Main/MenuAddModal/OptionButton.tsx
@@ -1,4 +1,4 @@
-import Button from "../Button";
+import Button from "../../Button";
 import styles from "./OptionButton.module.css";
 
 type OptionButtonProps = {
